requirements.txt
openai>=2.8.0
python-dotenv>=1.0.0
boto3>=1.35.0
requests>=2.32.0
